Backend/requirements.txt
Django==4.1.3
djangorestframework==3.13.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.13.0
Pillow==9.3.0
django-cleanup==6.0.0

